// src/app/shared/model/userorderhistory-detail.model.ts
export class UserOrderHistoryDetail
{
    userOrderId: string='';
    address: string='';
    price: string='';
    orderDate: string='';
    canCancelOrder: boolean=true;
    currentDate: string='';
    deliveryDate: string='';
    deliveryStatus: string='';               
}

export class UserOrderDetails
{
    userOrderId: string='';
    title: string='';
    color: string='';
    size: string='';    
    quantity: string='';
    price: string='';            
}